+++ src/BOOSOP.h
@@ -47,7 +47,7 @@ template<typename distribution_type> class BOOSOP {
      */
     std::tuple<std::vector<double>, std::vector<data::Quaternion>>
     compute(const float* distances,
-            const double* weights,
+            const float* weights,
             const int* num_neighbors,
             size_t N_particles,
             const unsigned int m,
@@ -55,7 +55,7 @@ template<typename distribution_type> class BOOSOP {
             size_t ylms_shape_0,
             const float* quad_positions,
             size_t quad_positions_shape_0,
-            const double* quad_weights) const;
+            const float* quad_weights) const;
 
     /**
      * @brief Compute BOOSOP at given rotations for each point.
@@ -78,8 +78,8 @@ template<typename distribution_type> class BOOSOP {
      *
      */
     std::vector<double> refine(const float* distances,
-                               const double* rotations,
-                               const double* weights,
+                               const float* rotations,
+                               const float* weights,
                                const int* num_neighbors,
                                size_t N_particles,
                                const unsigned int m,
@@ -87,7 +87,7 @@ template<typename distribution_type> class BOOSOP {
                                size_t ylms_shape_0,
                                const float* quad_positions,
                                size_t quad_positions_shape_0,
-                               const double* quad_weights) const;
+                               const float* quad_weights) const;
 
     private:
     /**
