+++ src/PGOP.h
@@ -41,9 +41,9 @@ class PGOP {
      */
     std::tuple<std::vector<double>, std::vector<data::Quaternion>>
     compute(const float* distances,
-            const double* weights,
+            const float* weights,
             const int* num_neighbors,
-            const double* sigmas,
+            const float* sigmas,
             const size_t N_particles_in_neighbors) const;
 
     private:
